+++ lib/src/models/em_deviceInfo.dart
@@ -1,8 +1,4 @@
-
-
 class EMDeviceInfo {
-
-
   EMDeviceInfo._private();
 
   Map toJson() {
@@ -22,8 +18,10 @@ class EMDeviceInfo {
 
   /// 设备资源描述
   String? get resource => _resource;
+
   /// 设备的UUID
   String? get deviceUUID => _deviceUUID;
+
   /// 设备名称
   String? get deviceName => _deviceName;
